Update typing of 'fake_lib'. Code edits.
