+++ fake_lib/__init__.py
@@ -1,0 +1,5 @@
+from fake_factory import FakeFactory
+import providers
+
+
+__all__ = ['FakeFactory', 'providers']
